requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
numpy
audio-recorder-streamlit
fpdf